新闻咨询详情页使用mvvm重构
