+++ qyb/src/main/res/layout/activity_news_item.xml
@@ -1,19 +1,23 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical"
-    >
+<layout xmlns:android="http://schemas.android.com/apk/res/android">
 
-    <include layout="@layout/toolbar"/>
+    <data>
 
-<!--<FrameLayout-->
-    <!--android:layout_width="fill_parent"-->
-    <!--android:layout_height="50dp"-->
-    <!--android:background="#4CAF50"-->
-    <!-->-->
-    <!--<TextView-->
+    </data>
+
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:orientation="vertical">
+
+        <include layout="@layout/toolbar" />
+
+        <!--<FrameLayout-->
+        <!--android:layout_width="fill_parent"-->
+        <!--android:layout_height="50dp"-->
+        <!--android:background="#4CAF50"-->
+        <!-->-->
+        <!--<TextView-->
         <!--android:padding="7dp"-->
         <!--android:layout_width="fill_parent"-->
         <!--android:layout_height="fill_parent"-->
@@ -22,60 +26,59 @@
         <!--style="@style/textSytle"-->
         <!--android:gravity="center"-->
         <!--/>-->
-    <!--</FrameLayout>-->
+        <!--</FrameLayout>-->
 
-    <RelativeLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        >
+        <RelativeLayout
+            android:layout_width="match_parent"
+            android:layout_height="match_parent">
 
-        <TextView
-            android:id="@+id/title"
-            android:padding="7dp"
-            android:layout_width="fill_parent"
-            android:layout_height="wrap_content"
-            android:text=""
-            android:textSize="18sp"
-            android:gravity="center"
-            />
-        <LinearLayout
-            android:paddingTop="10dp"
-            android:id="@+id/myimage"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:gravity="center"
-            android:layout_centerHorizontal="true"
-            >
+            <TextView
+                android:id="@+id/title"
+                android:layout_width="fill_parent"
+                android:layout_height="wrap_content"
+                android:gravity="center"
+                android:padding="7dp"
+                android:text=""
+                android:textSize="18sp" />
+
+            <LinearLayout
+                android:id="@+id/myimage"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_centerHorizontal="true"
+                android:gravity="center"
+                android:paddingTop="10dp">
 
 
-        <ImageView
-            android:id="@+id/image1"
-            android:layout_width="150dp"
-            android:layout_height="200dp"
-            android:paddingRight="5dp"
-            android:src="@drawable/jwc_news_logo"
-            />
-        <ImageView
-            android:paddingLeft="5dp"
-            android:id="@+id/image2"
-            android:layout_toRightOf="@+id/image1"
-            android:layout_width="150dp"
-            android:layout_height="200dp"
-            android:src="@drawable/jwc_news_logo"
-            />
-        </LinearLayout>
-        <ScrollView
-            android:layout_below="@+id/myimage"
-            android:layout_centerInParent="true"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content">
-        <TextView
-            android:padding="20dp"
-            android:id="@+id/content"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="oh no~~ ，新闻内容没找到"
-            />
-        </ScrollView>
+                <ImageView
+                    android:id="@+id/image1"
+                    android:layout_width="150dp"
+                    android:layout_height="200dp"
+                    android:paddingRight="5dp"
+                    android:src="@drawable/jwc_news_logo" />
+
+                <ImageView
+                    android:id="@+id/image2"
+                    android:layout_width="150dp"
+                    android:layout_height="200dp"
+                    android:layout_toRightOf="@+id/image1"
+                    android:paddingLeft="5dp"
+                    android:src="@drawable/jwc_news_logo" />
+            </LinearLayout>
+
+            <ScrollView
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_below="@+id/myimage"
+                android:layout_centerInParent="true">
+
+                <TextView
+                    android:id="@+id/content"
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:padding="20dp"
+                    android:text="oh no~~ ，新闻内容没找到" />
+            </ScrollView>
         </RelativeLayout>
-</LinearLayout>
+    </LinearLayout>
+</layout>
